+++ api/src/routes/categoriesCRUD/deleteCategorie.ts
@@ -3,11 +3,20 @@ const {Category} = require('../../db');
 export async function deleteCategorie (req: any, res: any) {
     try {
         const { name } = req.params;
-        Category.destroy({
-            where: {"name" : name}
-        })
-        return res.status(200).send(`The category ${name} has been deleted`);
-    } catch (err) {
-        return res.status(404).send(err);
+        if(name){
+            Category.destroy({
+                where: {"name" : name}
+            })
+            return res.status(200).send(`The category ${name} has been deleted`);
+        } else {
+            return res.status(404).send("The name has not been recognized or has not been entered, please try again.")
+        }
+    } catch (err: any) {
+        const errName = err.name;
+        const errCode = err.code;
+        const errMessage = err.message;
+        return res.status(404).send(errName ? 
+            `Error ${errCode}: ${errName} - ${errMessage}` : 
+            "Something went wrong, please try again.");
     }
 }
